requirements.txt
pandas
numpy
scikit-learn
mlflow
notebook
python-box
pyYAML
tqdm
ensure
joblib
types-pyYAML
Flask
Flask-Cors
python-dotenv